+++ app.js
@@ -2,42 +2,19 @@
 require('dotenv').config();
 const path = require('path');
 const express = require('express');
-const { celebrate, Joi } = require('celebrate');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const mongoose = require('mongoose');
 const { errors } = require('celebrate');
 const cors = require('cors');
+const limiter = require('./middlewares/rateLimiter');
+const errorHandler = require('./middlewares/errorHandler');
+const { createUserValidation, loginValidation } = require('./middlewares/validations');
 const { login, createUser } = require('./controllers/users');
 const authMiddleware = require('./middlewares/auth');
 const NotFoundError = require('./errors/not-found-error');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const { DB_ADDRESS } = require('./config');
-
-const errorHandler = (err, req, res, next) => {
-  console.log(err.status);
-  const statusCode = err.status || 500;
-  const message = statusCode === 500 ? 'На сервере произошла ошибка' : err.message;
-
-  res.status(statusCode).send({ message });
-
-  next();
-};
-
-const createUserValidation = celebrate({
-  body: Joi.object().keys({
-    name: Joi.string().min(2).max(30),
-    email: Joi.string().email().required(),
-    password: Joi.string().required(),
-  }),
-});
-
-const loginValidation = celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().email().required(),
-    password: Joi.string().required(),
-  }),
-});
 
 const { PORT = 3000 } = process.env;
 const app = express();
@@ -64,9 +41,12 @@ app.use(cookieParser());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(requestLogger);
-
+app.use(limiter);
 app.post('/signin', loginValidation, login);
 app.post('/signup', createUserValidation, createUser);
+app.post('/signout', (req, res) => {
+  res.clearCookie('jwt').send({ message: 'Пользователь вышел' });
+});
 app.use(authMiddleware);
 app.use('/users', require('./routes/users'));
 app.use('/movies', require('./routes/movies'));
