+++ routes/movies.js
@@ -1,31 +1,9 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
+const { createMovieValidation, deleteMovieValidation } = require('../middlewares/validations');
 
 const {
   getMovies, createMovie, deleteMovieById,
 } = require('../controllers/movies');
-
-const createMovieValidation = celebrate({
-  body: Joi.object().keys({
-    country: Joi.string().required(),
-    director: Joi.string().required(),
-    duration: Joi.number().required(),
-    year: Joi.string().required(),
-    description: Joi.string().required(),
-    image: Joi.string().uri().required(),
-    trailerLink: Joi.string().uri().required(),
-    nameRU: Joi.string().required(),
-    nameEN: Joi.string().required(),
-    thumbnail: Joi.string().uri().required(),
-    movieId: Joi.number().required(),
-  }),
-});
-
-const deleteMovieValidation = celebrate({
-  params: Joi.object().keys({
-    movieId: Joi.string().hex().length(24).required(),
-  }),
-});
 
 router.get('/', getMovies);
 router.post('/', createMovieValidation, createMovie);
